refactor: regroup internals into methods/ and performance/ (Phase 2)
Phase 2 of the architectural revision (ARCHITECTURE.md §8). Internal
reorganization only — the public API is unchanged: `import freqprob` and
every `freqprob.<Name>` continue to work, backed by re-exports in
__init__.py.

- Estimators grouped by family under freqprob/methods/:
  baselines (Uniform/Random/MLE), additive (Lidstone/Laplace/ELE),
  goodturing (WittenBell/SimpleGoodTuring), certainty (CertaintyDegree),
  kneser_ney (KneserNey/ModifiedKneserNey), interpolated, bayesian.
- Performance infrastructure grouped under freqprob/performance/:
  lazy, vectorized, streaming, memory_efficient, profiling.
- Split the utils.py junk drawer into metrics.py (perplexity, cross_entropy,
  kl_divergence, model_comparison) and text.py (generate_ngrams,
  word_frequency, ngram_frequency).
- base.py and cache.py stay top-level as shared foundations (methods depend
  on cache, so it is not buried under performance/).
- Cross-package imports use absolute form (from freqprob.base import ...) to
  satisfy the flake8-tidy-imports (TID) rule.
- Update in-repo test imports of internal module paths to the new locations;
  public classes are imported from the top-level namespace.

Verified: public API intact (41 names), wheel packages all subpackages with
py.typed, ruff/mypy/bandit clean, 229 passed / 5 skipped, 83.69% coverage.
