+++ src/freqprob/methods/__init__.py
@@ -1,0 +1,1 @@
+"""Smoothing method families (grouped for maintainability; import from the top-level ``freqprob`` namespace)."""
