+++ src/freqprob/methods/goodturing.py
@@ -1,22 +1,15 @@
-"""Advanced probability scoring methods.
-
-This module implements sophisticated smoothing methods including
-Witten-Bell discounting, Certainty Degree estimation, and
-Simple Good-Turing smoothing. These methods use more complex
-statistical techniques to estimate probability distributions.
-"""
+"""Good-Turing-family smoothing: Witten-Bell discounting and Simple Good-Turing."""
 
 import math
 from dataclasses import dataclass
 
-# Import 3rd-party modules
 import numpy as np
 import scipy
 import scipy.linalg
 import scipy.stats
 
-from .base import FrequencyDistribution, Probability, ScoringMethod, ScoringMethodConfig
-from .cache import cached_computation, cached_sgt_computation
+from freqprob.base import FrequencyDistribution, ScoringMethod, ScoringMethodConfig
+from freqprob.cache import cached_computation, cached_sgt_computation
 
 
 @dataclass
@@ -34,54 +27,6 @@ class WittenBellConfig(ScoringMethodConfig):
     bins: int | None = None
 
     logprob: bool = True
-
-
-@dataclass
-class CertaintyDegreeConfig(ScoringMethodConfig):
-    """Configuration for Certainty Degree estimation.
-
-    Attributes:
-    ----------
-    bins : int | None
-        Total number of possible bins/elements (default: vocabulary size)
-    unobs_prob : Probability
-        Reserved probability mass for unobserved elements (default: 0.0)
-    logprob : bool
-        Whether to return log-probabilities (default: True)
-    """
-
-    bins: int | None = None
-
-    unobs_prob: Probability = 0.0
-    logprob: bool = True
-
-
-@dataclass
-class SimpleGoodTuringConfig(ScoringMethodConfig):
-    """Configuration for Simple Good-Turing smoothing.
-
-    Attributes:
-    ----------
-    p_value : float
-        Confidence level for smoothing threshold (default: 0.05)
-    default_p0 : float | None
-        Fallback probability for unobserved elements (default: None)
-    bins : int | None
-        Total vocabulary size for per-word probability calculation (default: None).
-        If None, uses heuristic: bins = V_observed + N1 (singletons).
-        This determines how p0 (total unseen mass) is distributed across unseen words.
-    logprob : bool
-        Whether to return log-probabilities (default: True)
-    allow_fail : bool
-        Whether to raise errors on invalid assumptions (default: True)
-    """
-
-    p_value: float = 0.05
-
-    default_p0: float | None = None
-    bins: int | None = None
-    logprob: bool = True
-    allow_fail: bool = True
 
 
 class WittenBell(ScoringMethod):
@@ -152,89 +97,32 @@ class WittenBell(ScoringMethod):
                 self._unobs = t / (z * (n + t))
 
 
-class CertaintyDegree(ScoringMethod):
-    """Returns a Certainty Degree probability distribution.
+@dataclass
+class SimpleGoodTuringConfig(ScoringMethodConfig):
+    """Configuration for Simple Good-Turing smoothing.
 
-    In this distribution a mass probability is reserved for unobserved samples
-    from a computation of the degree of certainty that the are no unobserved
-    samples.
-
-    Under development and test by Tiago Tresoldi, this is an experimental
-    probability distribution that should not be used as the sole or main
-    distribution for the time being.
-
-    Parameters
+    Attributes:
     ----------
-    freqdist : dict
-        Frequency distribution of samples (keys) and counts (values) from
-        which the probability distribution will be calculated.
-    bins: int
-        The optional number of sample bins that can be generated by the
-        experiment that is described by the probability distribution. If not
-        specified, it will default to the number of samples in the frequency
-        distribution.
-    unobs_prob : float
-        An optional mass probability to be reserved for unobserved states,
-        from 0.0 to 1.0. If not specified, it will default to 0.0.
+    p_value : float
+        Confidence level for smoothing threshold (default: 0.05)
+    default_p0 : float | None
+        Fallback probability for unobserved elements (default: None)
+    bins : int | None
+        Total vocabulary size for per-word probability calculation (default: None).
+        If None, uses heuristic: bins = V_observed + N1 (singletons).
+        This determines how p0 (total unseen mass) is distributed across unseen words.
     logprob : bool
-        Whether to return the log-probabilities (default) or the
-        probabilities themselves. When using the log-probabilities, the
-        counts are automatically corrected to avoid domain errors.
+        Whether to return log-probabilities (default: True)
+    allow_fail : bool
+        Whether to raise errors on invalid assumptions (default: True)
     """
 
-    __slots__ = ()
+    p_value: float = 0.05
 
-    def __init__(
-        self,
-        freqdist: FrequencyDistribution,
-        bins: int | None = None,
-        unobs_prob: Probability = 0.0,
-        logprob: bool = True,
-    ) -> None:
-        """Initialize Certainty Degree estimation."""
-        config = CertaintyDegreeConfig(bins=bins, unobs_prob=unobs_prob, logprob=logprob)
-        super().__init__(config)
-        self.name = "Certainty Degree"
-        self.fit(freqdist)
-
-    @cached_computation()
-    def _compute_probabilities(self, freqdist: FrequencyDistribution) -> None:
-        """Compute Certainty Degree probabilities."""
-        bins = self.config.bins
-
-        unobs_prob = self.config.unobs_prob
-
-        # Obtain the parameters for probability calculation.
-        n = sum(freqdist.values())
-        b = len(freqdist)
-        z = bins or b
-
-        # Calculate the mass of probability space to reserve and use this value to
-        # correct the Maximum-Likelihood Estimate for each sample.
-        # Note that, for very large values of N, this will underflow because we
-        # effectively have a large confidence of having observed all the
-        # samples that matter; this is a problem when taking the
-        # log-probability, as we'll ultimately raise a math domain error by
-        # asking for the logarithm of what is machine-represented as zero;
-        # for this reason, we take as the probability space the minimum value
-        # between 1.0 discounted the calculated mass and 1.0 discounted the
-        # minimum mass probability reserved.
-        if self.logprob:
-            # Ensure unobs_prob is not None and handle self._unobs initialization
-            unobs_prob = unobs_prob or 0.0
-            current_unobs = getattr(self, "_unobs", 0.0) or 0.0
-            unobs_prob = max(unobs_prob, current_unobs)
-            prob_space = min(1.0 - (b / (z + 1)) ** n, 1.0 - unobs_prob)
-            self._prob = {
-                sample: math.log((count / n) * prob_space) for sample, count in freqdist.items()
-            }
-            self._unobs = math.log(-(prob_space - 1.0))
-        else:
-            # Ensure unobs_prob is not None
-            unobs_prob = unobs_prob or 0.0
-            prob_space = min(1.0 - (b / (z + 1)) ** n, 1.0 - unobs_prob)
-            self._prob = {sample: (count / n) * prob_space for sample, count in freqdist.items()}
-            self._unobs = -(prob_space - 1.0)
+    default_p0: float | None = None
+    bins: int | None = None
+    logprob: bool = True
+    allow_fail: bool = True
 
 
 class SimpleGoodTuring(ScoringMethod):
