+++ src/freqprob/performance/__init__.py
@@ -1,0 +1,1 @@
+"""Performance infrastructure: lazy, streaming, vectorized, memory-efficient, profiling."""
